requirements.txt
asgiref==3.8.1
Django==5.1.1
django-filter==23.1
django-tables2==2.7.0
djangorestframework==3.15.2
setuptools==73.0.1
sqlparse==0.5.1
tablib==3.6.1
wheel==0.44.0